introduce SingleThreadWorker
