+++ apollo-normalized-cache/build.gradle.kts
@@ -10,6 +10,7 @@ kotlin {
     val commonMain by getting {
       dependencies {
         api(project(":apollo-api"))
+        api(project(":apollo-mpp-utils"))
         api(project(":apollo-normalized-cache-api"))
         api(groovy.util.Eval.x(project, "x.dep.kotlin.coroutines"))
       }
